requirements.txt
# --- requirements.txt ---

# Core Data Handling
pandas>=2.0.0,<3.0.0
numpy>=1.20.0,<2.0.0
openpyxl>=3.0.0,<4.0.0 # Para ler .xlsx

# Machine Learning & Modeling
scikit-learn>=1.1.0,<2.0.0 # Para treino, HPO, métricas
joblib>=1.0.0,<2.0.0 # Para salvar/carregar modelos

# Web Scraping (para jogos futuros)
selenium>=4.10.0,<5.0.0

# Interface (se usar)
rich>=13.0.0,<14.0.0 # Para CLI (dashboard.py)
# tkinter (GUI) é built-in

# Configuração
python-dotenv>=1.0.0,<2.0.0

# Utilitários
tqdm>=4.60.0,<5.0.0